+++ config/src/remaps/fixture.rs
@@ -1,6 +1,6 @@
-use crate::{schema::SchemaFixture, ConfigParseError};
-
-use super::schema::SchemaFixtureType;
+use crate::{
+    schema::fixture::FixtureSchema, schema::fixture_type::FixtureTypeSchema, ConfigParseError,
+};
 use std::collections::HashMap;
 
 #[derive(Debug)]
@@ -14,11 +14,11 @@ impl Fixture {
     pub(super) fn new(
         identifier: &str,
         start_addr: u16,
-        fixture_type: &SchemaFixtureType,
+        fixture_type: &FixtureTypeSchema,
     ) -> Fixture {
         let mut channels: HashMap<String, u16> = HashMap::new();
-        for (i, channel) in fixture_type.channels.iter().enumerate() {
-            channels.insert(channel.name.clone(), start_addr + i as u16);
+        for (i, channel) in fixture_type.get_channels().iter().enumerate() {
+            channels.insert(channel.get_name().to_string(), start_addr + i as u16);
         }
         Fixture {
             identifier: identifier.to_string(),
@@ -38,30 +38,34 @@ impl Fixture {
 }
 
 pub fn remap_fixtures(
-    fixtures: Vec<SchemaFixture>,
-    fixture_types: Vec<SchemaFixtureType>,
+    fixtures: &[FixtureSchema],
+    fixture_types: &[FixtureTypeSchema],
 ) -> Result<HashMap<String, Fixture>, ConfigParseError> {
-    let mut fixture_types_map: HashMap<String, SchemaFixtureType> = HashMap::new();
+    let mut fixture_types_map: HashMap<String, FixtureTypeSchema> = HashMap::new();
 
     for fixture_type in fixture_types {
-        fixture_types_map.insert(fixture_type.name.clone(), fixture_type);
+        fixture_types_map.insert(fixture_type.get_name().to_string(), fixture_type.clone());
     }
 
     let mut fixtures_map: HashMap<String, Fixture> = HashMap::new();
 
     for fixture in fixtures {
-        let fixture_type = fixture_types_map.get(&fixture.fixture_type);
+        let fixture_type = fixture_types_map.get(fixture.get_fixture_type_name());
 
         if fixture_type.is_none() {
             return Err(ConfigParseError::FixtureTypeNotFound(
-                fixture.fixture_type,
-                fixture.name,
+                fixture.get_fixture_type_name().to_string(),
+                fixture.get_name().to_string(),
             ));
         }
 
         fixtures_map.insert(
-            fixture.name.clone(),
-            Fixture::new(&fixture.name, fixture.start_addr, &fixture_type.unwrap()),
+            fixture.get_name().to_string(),
+            Fixture::new(
+                &fixture.get_name(),
+                *fixture.get_start_addr(),
+                &fixture_type.unwrap(),
+            ),
         );
     }