+++ config/src/test/parser.rs
@@ -1,55 +1,9 @@
-use crate::yaml::parse_yaml_into;
-
-const YAML: &str = r#"
-        config:
-          server:
-            binds: 0.0.0.0:3000
-            allow_direct_fixture_control: true
-          artnet:
-            binds: 0.0.0.0:6454
-            sends: 255.255.255.255:6454
-            broadcast: true
-            universe: 0
-        fixture_types:
-        - name: FixtureType1
-          channels:
-          - name: Channel1
-        fixtures:
-        - name: Fixture1
-          type: FixtureType1
-          start_addr: 1
-        bindings:
-        - identifier: Binding1
-          actions:
-          - Fixture1.Channel1: 1
-    "#;
-
-const YAML_ONLY_NECCESSARY: &str = r#"
-        config:
-          server:
-            binds: 0.0.0.0:3000
-          artnet:
-            binds: 0.0.0.0:6454
-            sends: 255.255.255.255:6454
-            broadcast: true
-            universe: 0
-        fixture_types:
-        - name: FixtureType1
-          channels:
-          - name: Channel1
-        fixtures:
-        - name: Fixture1
-          type: FixtureType1
-          start_addr: 1
-        bindings:
-        - identifier: Binding1
-          actions:
-          - Fixture1.Channel1: 1
-    "#;
+use super::common::{YAML, YAML_ONLY_NECCESSARY};
+use crate::{remaps::binding::StepsMode, yaml::parse_yaml, Binding};
 
 #[test]
 fn test_parse_fixtures() {
-    let result = parse_yaml_into(&YAML);
+    let result = parse_yaml(&YAML);
     assert!(result.is_ok());
 
     let (fixture_map, _, _) = result.unwrap();
@@ -70,24 +24,51 @@ fn test_parse_fixtures() {
 
 #[test]
 fn test_parse_bindings() {
-    let result = parse_yaml_into(&YAML);
+    let result = parse_yaml(&YAML);
     assert!(result.is_ok());
 
     let (_, bindings_map, _) = result.unwrap();
     assert!(bindings_map.get("Binding1").is_some());
-    assert_eq!(
-        bindings_map.get("Binding1").unwrap().get_identifier(),
-        "Binding1"
-    );
+    let binding = bindings_map.get("Binding1").unwrap();
+    match binding {
+        Binding::WithActions(bnd) => {
+            assert_eq!(bnd.get_identifier(), "Binding1")
+        }
+        _ => {
+            panic!("Binding2 is not a WithActions binding")
+        }
+    }
+}
 
-    let actions = bindings_map.get("Binding1").unwrap().get_actions();
-    assert_eq!(actions[0][0], "Fixture1.Channel1");
-    assert_eq!(actions[0][1], "1");
+#[test]
+fn test_parse_steps() {
+    let result = parse_yaml(&YAML);
+    assert!(result.is_ok());
+
+    let (_, bindings_map, _) = result.unwrap();
+    assert!(bindings_map.get("Binding2").is_some());
+    let binding = bindings_map.get("Binding2").unwrap();
+    match binding {
+        Binding::WithSteps(bnd) => {
+            assert_eq!(bnd.get_identifier(), "Binding2");
+            match bnd.get_mode() {
+                StepsMode::Once => {
+                    assert!(true)
+                }
+                _ => {
+                    panic!("Binding2 is not a Once binding")
+                }
+            }
+        }
+        _ => {
+            panic!("Binding2 is not a WithActions binding")
+        }
+    }
 }
 
 #[test]
 fn test_parse_config() {
-    let result = parse_yaml_into(&YAML);
+    let result = parse_yaml(&YAML);
     assert!(result.is_ok());
 
     let (_, _, config) = result.unwrap();
@@ -96,13 +77,15 @@ fn test_parse_config() {
     assert_eq!(config.get_artnet_send(), "255.255.255.255:6454");
     assert_eq!(config.get_artnet_universe(), 0);
     assert_eq!(config.get_allow_direct_fixture_control(), true);
+    assert_eq!(config.get_send_every_ms(), 100)
 }
 
 #[test]
 fn test_parse_config_only_necessary() {
-    let result = parse_yaml_into(&YAML_ONLY_NECCESSARY);
+    let result = parse_yaml(&YAML_ONLY_NECCESSARY);
     assert!(result.is_ok());
 
     let (_, _, config) = result.unwrap();
     assert_eq!(config.get_allow_direct_fixture_control(), false);
+    assert_eq!(config.get_send_every_ms(), 50);
 }